test: split dredd-test.js into multiple better themed files
